third_party/WebKit/LayoutTests/inspector/extensions/extensions-timeline-api-expected.txt

Tests timeline support in WebInspector Extensions API

 Started extension.
Running tests...
RUNNING TEST: extension_testTimeline
{
    onRecordingStarted : {
        addListener : <function>
        removeListener : <function>
    }
    onRecordingStopped : {
        addListener : <function>
        removeListener : <function>
    }
}
traceProvider.onRecordingStarted fired.
traceProvider.onRecordingStopped fired.
All tests done.
